+++ resources/views/purchase_order/index.blade.php
@@ -230,6 +230,10 @@
                 }).then(willDelete => {
                     if (willDelete) {
                         var href = $(this).attr('href');
+                        if (!href) {
+                            toastr.error(LANG.something_went_wrong);
+                            return;
+                        }
                         $.ajax({
                             method: 'DELETE',
                             url: href,
@@ -241,6 +245,12 @@
                                 } else {
                                     toastr.error(result.msg);
                                 }
+                            },
+                            error: function(xhr) {
+                                var msg =
+                                    (xhr.responseJSON && (xhr.responseJSON.msg || xhr.responseJSON.message)) ||
+                                    LANG.something_went_wrong;
+                                toastr.error(msg);
                             },
                         });
                     }
